fix: 修复日志过滤、窗口异常及代码格式
- 修复主进程日志过滤器误排除插件日志的问题
- 修复插件 widget 关闭时可能触发的 RuntimeError
- 修复历史表格 show_fields 使用 set 导致的列顺序不一致
- 将 OtherInfoBase 改为顶级导入以修复潜在的循环依赖
- 设置独立插件窗口提示栏固定高度
- 统一代码格式并重构历史插件 UI 组件
