+++ requirements.txt
@@ -7,3 +7,4 @@ pywin32>=311
 loguru>=0.7.3
 debugpy>=1.8.20
 PyQt5-stubs>=5.15.6.0
+requests>=2.33.1
